Move auth files to gpm
